refactor: move lib dir
