+++ lib/firebase/client.ts
@@ -1,7 +1,7 @@
 import firebase from 'firebase/app'
 import 'firebase/auth'
 import 'firebase/firestore'
-import clientCredential from '../../../credentials/client'
+import clientCredential from '../../credentials/client'
 
 const app = !firebase.apps.length
   ? firebase.initializeApp(clientCredential)
