+++ lib/firebase/type.ts
