+++ pages/api/type.ts
@@ -1,4 +1,4 @@
-import { App, DecodedIdToken } from '../../src/lib/firebase/type'
+import { App, DecodedIdToken } from '../../lib/firebase/type'
 import { NextApiRequest, NextApiResponse } from 'next'
 
 export interface Request extends NextApiRequest {
